Successfull sim! with kernel of MyStructs256
We finally got the simulation to fully work with a properly written dispersal kernel. Still working on the visualisation aspect of MyStructs256 and trying to merge cell rules with dispersal, which for some reason are causing NA's, but the solution must lie in the oscillaition of those values til infinity I think. Let's go!